add readme, update tests

+++ tests/integration/post.test.ts
@@ -1,32 +1,39 @@
-import { describe, it, expect, beforeAll, afterAll } from 'vitest';
+import { describe, it, expect, afterAll, beforeEach, beforeAll } from 'vitest';
 import request from 'supertest';
 import { app, prisma } from '../../src/server';
 
 beforeAll(async () => {
-  await prisma.post.deleteMany();
-  await prisma.user.deleteMany();
+  if (await prisma.user.count() <= 0 && !(await prisma.user.findFirst({ where: {id: 1} }))) {
+    await prisma.user.deleteMany();
 
-  await prisma.post.create({
-    data: {
-      id: 1,
-      title: 'Test Post',
-      content: 'Content of the test post',
-      createdAt: new Date('1990-01-01T00:00:00.000Z'),
-      userId: 1,
-    },
-  });
+    await prisma.user.create({
+      data: {
+        id: 1,
+        name: 'Test User',
+        email: 'test.email@gmail.com',
+        birthdate: '1990-01-01T00:00:00.000Z',
+      }
+    });
+  }
 
-  await prisma.user.create({
-    data: {
-      id: 1,
-      name: 'Test User',
-      email: 'test@example.com',
-      birthdate: new Date('1990-01-01T00:00:00.000Z'),
-    },
-  });
+  if (await prisma.post.count() <= 0 && !(await prisma.post.findFirst({ where: {id: 1} }))) {
+    await prisma.post.deleteMany();
+
+    await prisma.post.create({
+      data: {
+        id: 1,
+        title: 'Test Post',
+        content: 'This is a test post',
+        createdAt: '2021-01-01T00:00:00.000Z',
+        userId: 1,
+      },
+    });
+  }
 });
 
 afterAll(async () => {
+  await prisma.user.deleteMany();
+  await prisma.post.deleteMany();
   await prisma.$disconnect();
 });
 
@@ -35,10 +42,10 @@ describe('POST /posts', () => {
     const response = await request(app)
       .post('/posts')
       .send({
-        id: 1,
+        id: 2,
         title: 'New Post',
         content: 'Content of the new post',
-        createdAt: new Date('1990-01-01T00:00:00.000Z'),
+        createdAt: '1990-01-01T00:00:00.000Z',
         userId: 1,
       });
 
@@ -63,7 +70,7 @@ describe('PUT /posts/:id', () => {
       .send({
         title: 'Updated Post',
         content: 'Updated content of the post',
-        createdAt: new Date('1990-01-01T00:00:00.000Z'),
+        createdAt: '1990-01-01T00:00:00.000Z',
         userId: 1,
       });
 
@@ -95,8 +102,6 @@ describe('GET /posts', () => {
   it('should return a single post by id', async () => {
     const response = await request(app).get('/posts/1');
 
-    console.log(response.body);
-
     expect(response.status).toBe(200);
     expect(response.body).toHaveProperty('id', 1);
   });
@@ -106,7 +111,6 @@ describe('DELETE /posts/:id', () => {
   it('should delete a post', async () => {
     const response = await request(app).delete('/posts/1');
 
-    console.log(response.headers);
     expect(response.status).toBe(204);
   });
